+++ examples/tencentcloud-security-group/main.tf
@@ -1,4 +1,8 @@
 resource "tencentcloud_security_group" "group" {
   name        = "${var.short_name}"
   description = "New security group"
+
+  tags = {
+    "test" = "test"
+  }
 }