+++ CHANGELOG.md
@@ -7,6 +7,8 @@ FEATURES:
 ENHANCEMENTS:
 
 * Data Source: `tencentcloud_security_groups`: add optional argument `result_output_file` and new attributes `ingress`, `egress` for list `security_groups`.
+* Data Source: `tencentcloud_security_groups`: add optional argument `tags`.
+* Resource: `tencentcloud_security_group`: add optional argument `tags`.
 
 BUG FIXES:
 * Resource: `tencentcloud_clb_listener_rule`: fix unclear description and errors in example.